+++ src/components/index.js
@@ -11,3 +11,4 @@ export { default as FilterBar } from './filter-bar';
 export { default as DatePicker } from './date-picker';
 export { default as TimePicker } from './time-picker';
 export { default as DateTimePicker } from './date-time-picker';
+export { default as TimeRangePicker } from './time-range-picker';