+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- TimeRangePicker component by [@amit-y](https://github.com/amit-y)
+
 ## [1.20.0] - 2023-12-18
 
 ### Added